+++ AVACLI/commands/routes.js
@@ -1,5 +1,5 @@
-import { terminalPrefix, getRoutes } from "../CoreUtil";
 import Table from "cli-table";
+import { terminalPrefix, getRoutes } from "../../AVACore/CoreUtil";
 
 
 /**