+++ AVACLI/commands/update.js
@@ -1,6 +1,6 @@
 import fs from "fs";
-import * as CoreUtil from "../CoreUtil";
 import { exec } from "child_process";
+import * as CoreUtil from "../../AVACore/CoreUtil";
 
 
 /**